+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,0 +1,57 @@
+package kitchenpos.tablegroup.application;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.repository.OrderTableRepository;
+import kitchenpos.tablegroup.application.dto.TableGroupRequest;
+import kitchenpos.tablegroup.domain.OrderTables;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.repository.TableGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+@Service
+public class TableGroupService {
+    private final OrderTableRepository orderTableRepository;
+    private final TableGroupRepository tableGroupRepository;
+    private final TableGroupValidator tableGroupValidator;
+
+    public TableGroupService(final OrderTableRepository orderTableRepository,
+                             final TableGroupRepository tableGroupRepository,
+                             final TableGroupValidator tableGroupValidator) {
+        this.orderTableRepository = orderTableRepository;
+        this.tableGroupRepository = tableGroupRepository;
+        this.tableGroupValidator = tableGroupValidator;
+    }
+
+    @Transactional
+    public TableGroup create(final TableGroupRequest tableGroupRequest) {
+        final List<Long> orderTableIds = tableGroupRequest.getOrderTableIds();
+
+        if (CollectionUtils.isEmpty(orderTableIds) || orderTableIds.size() < 2) {
+            throw new IllegalArgumentException();
+        }
+
+        final TableGroup tableGroup = tableGroupRepository.save(new TableGroup(LocalDateTime.now()));
+        final OrderTables orderTables = new OrderTables(orderTableRepository.findAllByIdIn(orderTableIds));
+
+        orderTables.group(tableGroup.getId(), orderTableIds.size());
+
+        return tableGroupRepository.save(tableGroup);
+    }
+
+    @Transactional
+    public void ungroup(final Long tableGroupId) {
+        final OrderTables orderTables = new OrderTables(orderTableRepository.findAllByTableGroupId(tableGroupId));
+
+        final List<Long> orderTableIds = orderTables.getOrderTables().stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        tableGroupValidator.validateUngroup(orderTableIds);
+        orderTables.ungroup();
+    }
+}
