+++ src/main/java/kitchenpos/tablegroup/application/TableGroupValidator.java
@@ -1,0 +1,24 @@
+package kitchenpos.tablegroup.application;
+
+import java.util.Arrays;
+import java.util.List;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.repository.OrderRepository;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TableGroupValidator {
+
+    private final OrderRepository orderRepository;
+
+    public TableGroupValidator(final OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    public void validateUngroup(final List<Long> orderTableIds) {
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
