+++ src/main/java/kitchenpos/tablegroup/application/dto/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.tablegroup.application.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
