+++ src/main/java/kitchenpos/tablegroup/application/dto/TableIdRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.tablegroup.application.dto;
 
 public class TableIdRequest {
