+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,0 +1,39 @@
+package kitchenpos.tablegroup.domain;
+
+import java.time.LocalDateTime;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class TableGroup {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private LocalDateTime createdDate;
+
+    protected TableGroup() {
+    }
+
+    public TableGroup(final LocalDateTime createdDate) {
+        this.createdDate = createdDate;
+    }
+
+    public TableGroup(final Long id, final LocalDateTime createdDate) {
+        this.id = id;
+        this.createdDate = createdDate;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+}
