+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -4,14 +4,15 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.application.dto.TableGroupRequest;
-import kitchenpos.application.dto.TableIdRequest;
+import kitchenpos.tablegroup.application.dto.TableGroupRequest;
+import kitchenpos.tablegroup.application.dto.TableIdRequest;
 import kitchenpos.application.support.IntegrationTest;
-import kitchenpos.domain.MenuGroupRepository;
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.ordertable.domain.repository.OrderTableRepository;
+import kitchenpos.product.domain.repository.ProductRepository;
+import kitchenpos.tablegroup.application.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
